+++ app/Http/Controllers/OrderController.php
@@ -13,31 +13,25 @@ class OrderController extends Controller
     /**
      * Display a listing of the resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Http\JsonResponse
      */
     public function index()
     {
         $orders = Order::OrderBy('id','desc')->get();
 
-        return response()->json(['orders' => $orders]);
-    }
-
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function create()
-    {
-        //
+        return response()->json([
+            'orders' => $orders
+        ]);
     }
 
     /**
      * Store a newly created resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Http\JsonResponse
      */
+
+    // TODO: Create request file
     public function store(Request $request)
     {
         $input = $request->except('_token');
@@ -48,48 +42,34 @@ class OrderController extends Controller
             'car_id' => 'exists:cars,id',
             'user_id' => 'exists:users,id'
         ]);
-        if($validator->fails()){
-            return response()->json($validator->errors()->toJson(), 400);
+        if ($validator->fails()) {
+            return response()->json([
+                'error' => $validator->errors()->toJson()
+            ], 400);
         }
 
-        // $validated = $request->validated();
+        Order::create([
+            'quantity' => $request->quantity,
+            'item_price' => $request->item_price,
+            'car_id' => $request->car_id,
+            'user_id' => $request->user_id
+        ]);
 
-        $order = new Order();
-        $order->fill($input);
-        $order->save();
-
-        return response()->json(['message' => 'Order successfully created']);
+        return response()->json([
+            'message' => 'Order successfully created'
+        ]);
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function show($id)
-    {
-        //
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function edit($id)
-    {
-        //
-    }
 
     /**
      * Update the specified resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
      * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Http\JsonResponse
      */
+
+    // TODO: Create request file
     public function update(Request $request, $id)
     {
         $input = $request->except('_token','_method','id');
@@ -101,29 +81,54 @@ class OrderController extends Controller
             'user_id' => 'exists:users,id'
         ]);
         if($validator->fails()){
-            return response()->json($validator->errors()->toJson(), 400);
+            return response()->json([
+                'error' => $validator->errors()->toJson()
+            ], 400);
         }
 
-        // $validated = $request->validated();
+        $order = Order::find($id);
+        if ($order) {
+            $order->create([
+                'quantity' => $request->quantity,
+                'item_price' => $request->item_price,
+                'status_id' => $request->status_id,
+                'car_id' => $request->car_id,
+                'user_id' => $request->user_id
+            ]);
+        } else {
+            return response()->json([
+                'error' => 'brand does not exist'
+            ], 400);
+        }
 
-        $category = Order::find($id);
-        $category->fill($input);
-        $category->update();
-
-        return response()->json(['message' => 'Order successfully updated']);
+        return response()->json([
+            'message' => 'Order successfully updated'
+        ]);
     }
 
     /**
      * Remove the specified resource from storage.
      *
      * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Http\JsonResponse
      */
     public function destroy($id)
     {
         $destroy = Order::find($id);
-        $destroy->delete();
+        if ($destroy) {
+            return response()->json([
+                'message' => 'Order successfully deleted'
+            ],200);
 
-        return response()->json(['message' => 'Order successfully deleted']);
+            $destroy->delete();
+        } else {
+            return response()->json([
+                'error' => 'Order does not exist'
+            ], 400);
+        }
+
+        return response()->json([
+            'message' => 'Order successfully deleted'
+        ]);
     }
 }
